beaker-wizard: pass include_fields to getbug() call
Apparently the 'attachments' field is not included by default, we need
to explicitly request it.

With python-bugzilla 1.x this didn't matter, because its default
behaviour is to silently go back to Bugzilla with another XMLRPC call to
fill in any missing attributes when they are accessed.

But this "autorefresh" behaviour is disabled by default in
python-bugzilla 2.x, and is indeed a bad idea, so let's just explicitly
request the fields we know we need up front instead.

Bug: 1422385
Change-Id: I066be3400c56a86a1b8a66575c54cab0a89ca719