avoid extra COUNT() queries before every scheduling step
Instead of first counting rows, we can just execute the query itself and
then decide if we did any work based on whether we go through any
iterations of the loop.

This necessitates some tweaks to the logging, to prevent beakerd from
constantly spewing even when there is no scheduling activity being done.

Bug: 1426745
Change-Id: Ie67cec2f9fb7d3e6b9891b2d00a1e0297df22a8c