bkr-update-inventory: report human error message to user
update-inventory allows the user to specify N number of servers on
the command line to trigger an inventory update for. This means that
the update-inventory command needs to catch any errors received back
from the server so they don't cause the program to terminate early.

Therefore, we need to reuse the exception handling used in
Client/src/bkr/client/main.py so that we print the human readable
error message instead of just the HTTP status code and its textual
version.

Bug: 1335394
Change-Id: I58ae3a8c7670dde44e72bfc2dc846d014e30a763
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>