server update-inventory: always force correct hostname
When running the inventory update task the running task needs to
communicate back to the beaker server which machine entry the updated
inventory should be applied to. This is done by locally (on the machine
running the task) pulling the configured hostname. This however can
result in many issues, lets say the kickstart script is custom and so
overwrites /etc/hosts resulting in the hostname being reported always as
host1.example.com. Or the dhcp server being used does not issue
hostnames in sync with the entries in beaker.

All of these systems are outside the control of beaker, and can result
in the update inventory task failing simply because the task cannot tell
the server which machine entry the information should be associated with.

Oddly enough the inventory update task allows for someone to override
the hostname used with a task parameter. We change the beaker server
to always issue this task parameter, eliminating all these subtle issues
for how this task could fail.

Bug: 1438666
Change-Id: Ie0eaaa54611a1ac0ea452ad975d2d2adaade4c39
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>