watchdog: use lxml instead of xmltramp
xmltramp does not correctly handle input in the form of XML serialized
to unicode with non-ASCII characters, which is what Beaker gives back in
its XMLRPC APIs. Let's just replace the last remaining usages of
xmltramp with lxml, which does handle this, as we have done everywhere
else.

RIP aaronsw.

Bug: 1457606
Change-Id: Ib46d6a29d02b994f36d8c48fffd6f9d8dcb6ce45