use POST for system excluded families form
Currently this defaults to a GET request, which is not really correct
since it is mutating state. More importantly, now that we have the
"Exclude All" button, if there is a large number of families being
excluded this can very easily hit the query string size limit.

Make the form submission use POST instead to avoid this problem.

Change-Id: If2ddc6d4f6f32d9470c5ce2839c5e640573f36ac