rationalise Job.delete_jobs, Job._delete_criteria, job.delete_jobs
These three methods had some circular and redundant logic that made it
difficult to see exactly what was going on. If we inline the Job methods
into the XMLRPC handlers where they are called, it is a little more
obvious what each one is doing.

For example, it wasn't necessarily clear before that deleting an
already-deleted job or a running job would be silently ignored. That
behaviour is not necessarily ideal, but it's what we already have and
now it is a bit more obvious.

Also it was extra confusing because, in spite of their names, these
methods are actually dealing only with "soft" deletion
(the to_delete column and soft_delete() method) and not "purging"
(the deleted column and delete() method).

Change-Id: I95d92992b7f32c326d0133bb6d7f5d7b19b08755