clarify deleted vs. expired vs. purged for jobs
Currently beaker-log-delete is actually performing two separate
operations, even though they are both mixed together in the code in
a kind of mish-mash: it is deleting jobs which have expired according to
their retention settings, *and* it is purging log files for jobs which
have either expired or been explicitly deleted by a user.

Adding to the confusion is that both these are referred to as "deletion"
or some variation of that in various places (hence soft_delete,
to_delete, etc).

This patch splits beaker-log-delete's operation into two distinct steps,
with some corresponding changes to make the code and naming clearer:

* The job.deleted column is renamed to job.purged. This is actually the
  timestamp at which the job's results and log files were purged. The
  corresponding .delete() method is renamed to .purge().

* The job.to_delete column is renamed to job.deleted. Previously this
  was only set when a user explicitly deleted a job (or "marked a job to
  be deleted" hence its original name). The corresponding method was
  called .soft_delete() but can now simply be inlined for clarity.

* When beaker-log-delete runs it first finds any job which is not
  deleted, but which should be deleted according to the expiry rules,
  and marks it as deleted. It then finds any deleted jobs which are not
  purged yet, and purges them. This is the two separate phases mentioned
  above.

* As a consequence of this, we have a new invariant: if a job is purged
  it is also deleted. Previously this was not necessarily true, but it
  is fixed up for old jobs by the migration script.

  So the .is_deleted condition which previously had to check if *either*
  to_delete *or* deleted is set, can now be simplified to just check
  deleted.

Bug: 1337789
Change-Id: Iab60e7ffdba65fe896d1296eaf5af9dc341f5473