Move remaining time tracker from the view to model
Keeping the watchdog time tracker in the model means that the extend
reservation modal has access to the latest (correct) time remaining
rather than the value with which the server was last updated.

Separate watchdog time remaining view from the rest of recipe runtime.
Update time remaining display in the UI to be populated from the
RecipeWatchdogTimeRemainingView render() method, and update the test
accordingly to cater for changes to the HTML

Bug: 1358619
Change-Id: I071332e0828214d2936dba4b2b20bd45168beebc