flush before returning from XMLRPC handling code
A lot can go wrong in the final flush, which is where most database
operations actually happen -- including the lock wait timeouts we are
seeing in bug 1097861.

By putting an explicit flush inside this try: block, it should mean
that a more specific XMLRPC fault will be given back to the client,
instead of just failing the whole request with a 500.

And the timing claimed by the log message will now include all database
operations, which are currently not being counted.

Change-Id: Id91acb47b7140b45c0e744d1af1958aa1034db9e