pylint: enable attribute checking
Now that we are using SQLAlchemy declarative style everywhere, and
back_populates rather than backrefs, we can finally turn this on.

We have a few "abstract" base classes with methods referencing
attributes defined only on subclasses, so these need some tweaking to
satisfy pylint.

And some dynamic trickery in SQLAlchemy and Alembic still needs a Pylint
plugin to make it aware of all the names which are available at runtime.

Change-Id: I1d5423413a552455b236e7b85ee06d6346e75593