tests: fix timing issue in test_group_link_escapes_uri_characters
The test case was doing a search for the group, before finding the name
and clicking it. But the group is *already* present on the page before
the search, and so WebDriver might find the old element which then
disappears before it can click it.

We can just drop the search step from the test, since the group will
always be present on the page.

Change-Id: I915e8633c58677042b396fefc2a5cc91d027fcd2