specify --hostname in kickstart
This is essentially working around a change in recent Anaconda versions
(bug 1290858).

In the past, Anaconda would just activate the one "installation" network
device, according to the various rules with ksdevice=. And then it would
set the system hostname based on the DHCP response for that network
device. On systems with multiple NICs, this was enough to ensure the
hostname was statically configured to the correct value, regardless
which NICs get activated in which order.

However now Anaconda no longer configures the system hostname. On each
reboot, the hostname will be dynamically set based on the DHCP response
for whichever NIC comes up first. But if there are multiple NICs it
could be random which one will win. And therefore the hostname can end
up set to some other unexpected value.

We can effectively restore the old behaviour by telling Anaconda what we
know the hostname *should* be, so that it will configure that on the
system, instead of letting it be non-deterministically set from DHCP.

Bug: 1346068
Change-Id: I0f6071e0ee4a8ba401d45f33ddf427cdb813e8dc