Added liveimg support to kickstart templates
Anaconda installs some OSs directly from an image (tarball or squashfs)
that is packed inside the ISO or compose, using the `liveimg` directive.

This patch presents two ways to install from a liveimg, one from a
direct image file and the other from an rpm that contains an image.

The following metadata variables were added to the kickstart templates:

liveimg_file - a relative path to the image itself
liveimg_rpm - a relative path to an rpm that ships an image
liveimg_post - additional line(s) to add to the %post section (usually
post processing is needed after installing with a liveimg)

Signed-off-by: Yuval Turgeman <yturgema@redhat.com>