tests: fix test_empty_and_element
So it always has a system. Right now if you run this test on its
own, it fails because there are no candidate systems and the recipe
aborts.The test passes as part of the whole suite because other tests
leave behind suitable systems.

Presumably it used to work because of create_test_env() but now we
don't call that. We can just make a system in the test itself so that
there is always at least one candidate.

create_test_env() method has been removed as it is no longer used.

Change-Id: Ifc9a0a8c6d763adf7c170f2b07ddcc14cc75af46