accept non-ASCII names in testinfo.desc
RFC5335 updated RFC2822 to accept UTF-8 in the name portion, so let's do
the same. We can cheat and just turn on re.UNICODE which makes \w match
any "alphanumeric" Unicode character in any script, which is probably
the most reasonable interpretation of the spec.

Matches the corresponding commit in rhts.

Bug: 1491658
Change-Id: I89400c78dee2bd93a025d6ed67652ce35b512427