Use fully qualified hostname for openstack instance
Commit 5553619f changed the hostname for recipes running in OpenStack
to "beaker-recipe-<id>" but this is not a fully qualified domain name.
Any Beaker tasks which are assuming the hostname is fully qualified
would therefore fail.

We noticed this in Beaker's own test suite, which started to fail because
unqualified hostnames are not valid in the domain name portion of
an email address.

Bug: 1506430
Change-Id: I3a7f64acb29e90107c9b718002e1b1f552733c2f