Added liveimg support to kickstart templates
Anaconda installs some OSs directly from an image (tarball or squashfs)
that is packed inside the ISO or compose, using the `liveimg` directive.

This patch presents two ways to install from a liveimg, one from a
direct image file and the other from an rpm that contains an image.

The following metadata variable was added to the kickstart templates:

* liveimg - a relative path to the squashfs file or rpm

Bug: 1420600
Change-Id: I5c8ad09f3882c13560aaa67fc784b332f0458974
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>