Added liveimg support to kickstart templates
Anaconda installs some OSs directly from an image (tarball or squashfs)
that is packed inside the ISO or compose, using the `liveimg` directive.

This patch presents two ways to install from a liveimg, one from a
direct image file and the other from an rpm that contains an image.

The following metadata variable was added to the kickstart templates:

* liveimg - a relative path to the squashfs file or rpm

Change-Id: I692d28605bb57c0d0c800f0c4aca59310e1bb465
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>