Make kickstart templates use installation columns
Previously the kickstart templates were not reliant on the installation
column in the recipe, but rather on recipe.distro_tree (or the specified
distro_tree).

Updated this to now use the installation column in recipe, except for
the case where distro tree is specifically passed for the
create_kickstart command.

Made kickstart generation work for the user-defined case when
distro_tree.id is None

Bug: 911515
Change-Id: I3e2717d65285fe17fe647ccf622973b17e3e1537