Add <distro/> tag option for specifying distro metadata
Previously the only way to specify a distro for provisioning was using
the <distroRequires/> tag. This allows for the user to specify the
distro metadata using the <distro/> tag. The user must either specify
the <distroRequires/> tag or <distro/> tag in order for the job XML to
be valid.

Make job results XML use new installation table. installation.distro_tree_id
is nullable for the user-defined scenario.

Bug: 911515
Change-Id: I890a6e7655f384a124054cf2f939d4df589ac4dd