track task "exclusive" releases in addition to excluded releases
The testinfo.desc metadata file in a task can have a Releases field,
indicating which operating system releases (OS majors, in Beaker
parlance) the task is applicable to. The task can be either applicable
to all releases except for certain excluded releases:

    Releases: -RedHatEnterpriseLinux4

or it can be exclusive to certain releases:

    Releases: RedHatEnterpriseLinux4

Previously, the latter case would be represented in Beaker's database as
excluding all *known* OS majors in the database *except* for the ones
listed in the field. However this meant when a new OS major appears that
was previously unknown, it would not be excluded for such tasks even
though it should be.

This introduces a separate table and relationship for tracking
"exclusive osmajors" for a task.

Bug: 800455
Change-Id: Ibb40dc3930f7a6d6e7f367e7b33be487fe913c9b