schedule newly released systems against recipes
... instead of constantly polling for queued recipes that have free
systems to run against.

This removes the schedule_queued_recipes() scheduling routine, in favour
of a new schedule_pending_systems(). A new column
system.scheduler_status is used to keep track of systems which have been
released but not yet checked against the recipe queue. The new
schedule_pending_systems() routine uses this column to check each newly
released system against the queued recipes in the database to see if any
of them can be assigned to the system.

As a consequence, when the scheduler sees a new job it now always takes
it through to Scheduled (if there is a free system already) or leaves it
in Queued until the new schedule_pending_systems() routine picks it out
of the queue.

Bug: 1519589
Change-Id: If07c7247d5e5e1f27dc3b69428f82fd170fe5115