fix OpenStack stuff to use public URLs instead of admin URLs
In our OpenStack environment, the "admin" endpoints exposed by the
Keystone service catalog used to work, but they are now restricted by
firewall rules (as they are designed to be). We can pass
interface='public' to make keystoneclient use the "public" endpoints
which are intended for regular clients.

... Or at least that's the theory. On keystoneclient 0.11 (RHEL6) it
ignores all of that and unconditionally tries to contact the admin
endpoint, so we have to make a raw HTTP request directly.

This also necessitated porting create_keystone_trust() to the new(-er)
usage style with explicitly constructed Auth, Session, and Client
objects. Although apparently that is already deprecated in favour of
some other thing called keystoneauth1.

Change-Id: Ib4fdb88150d154f3c44036267f99b2e622f25f0b