tests: ensure OpenStack test recipes go to OpenStack
There was a sort of hidden behaviour change in commit
a713fd08c864ddc8c26d33a043dcad90788ba9f6, namely that newly submitted
recipes will now always go to physical systems (if there are some free)
where before they would preferentially go to OpenStack if possible.

It remains to be seen if that's a desirable change or not, but for now
this commit just fixes the OpenStack tests so they continue to pass in
light of this change.

We didn't notice this problem when that commit was originally merged
because unfortunately the OpenStack tests were broken at the time.

Bug: 1519589
Change-Id: Ia5f56ffab168eabadba291a1033a20db1f520b33