make task columns non-NULLable
This effectively means it is no longer possible to call
Task.lazy_create().

For the real TaskLibrary.update_tasks() code, we can just switch to
a lookup-else-create pattern, which is not subject to race conditions
because the whole thing is enclosed by a global lock.

However there was a bunch of test code which was relying on the
lazy_create() behaviour which needed to be updated.

Bug: 859785
Change-Id: Ia2b955a6a5c413f0a1f9b9193c14d77c995c0a37