avoid CONSTRAINT violations for NULL on Boolean columns
On databases which support CHECK constraints, SQLAlchemy will generate
a constraint for Boolean columns to ensure their values are either 0 or
1. Like this:

    CONSTRAINT `CONSTRAINT_1` CHECK (`destructive` in (0,1))

Mariadb since version 10.2 will enforce these constraints.

However SQLAlchemy incorrectly generates such constraints even for
NULLable columns. But a NULL value will not satisfy the constraint.
Leading to an error like this when inserting:

    ERROR 4025 (23000): CONSTRAINT `CONSTRAINT_1` failed for `beaker_test`.`task`

Or if MySQL-python doesn't understand this new error code, it appears in
Python land as:

    InterfaceError: (_mysql_exceptions.InterfaceError) (-1, 'error totally whack')

Arguably these columns should be non-NULLable with a default of 0. And
SQLAlchemy should not generate unsatisfiable CHECK constraints. But as
a short term fix we can just disable the CHECK constraint on the
affected columns.

Change-Id: Ib3b13b52fbe0485e76157dad76260f93374b60be