Fix for recipe page when installation is null
When a recipe is either old and cancelled or aborted prior to being
scheduled, or queued at the time of migration (and therefore
recipe.provision() has not been called yet) there is no installation
row, and the recipe model installation attribute is null as a result.

Since this case has not been covered in the changes for BZ911515, a
condition for installation being null has been added to
recipe-installation.html and recipe-summary.html.

Bug: 911515
Change-Id: I8b18e1c6a98e28bfb1ae35d11ea9d39851f082b7