Fix distro XML generation for custom distro
There was previously an error in get_recipe method in proxy.py which
caused a failure in the labcontroller. This was due to a bug in
recipe.to_xml, in which the XML generation would fail if the optional
attributes were None (which it could be if the user did not provide it)

This fix ensured when these sub elements in the distro element are not
provided, XML generation of the submitted job should not include them,
and generation of XML should not fail.

Bug: 911515
Change-Id: I262df03fbd80d221fa8b0a9309ef241657387b7f