tests: Don't clobber STDOUT with STDERR output
Test-dependent Beaker applications such as the watchdog (beaker-watchdog) or the
hub (beaker-proxy) are being started as separate processes on test setup. The
programs run in foreground with debug level logging and the processes are setup
with redirecting all logging from STDERR to STDOUT.

The effect is, that during debugging sessions debug output from the in parallel
running programs are clobbering the debug shell, which makes it really hard to
step through the test code.

This patch changes how we create separate test processes and removes the
redirect of STDERR to STDOUT to make debugging easier.

It also changes the runtests.sh script in the integration tests script to
redirect any STDERR traffic into a log file under /tmp.

Change-Id: I219f3c7a4a3f4b2864b61fae9b914d7e27a5e6e4