tests: Don't clobber pdb prompts with logs from subprocesses
Test-dependent Beaker applications such as the watchdog (beaker-watchdog) or the
hub (beaker-proxy) are being started as separate processes on test setup. The
programs run in foreground with debug level logging and the processes are setup
with redirecting all logging from STDERR to STDOUT.

The effect is, that during debugging sessions debug output from the in parallel
running programs are clobbering the debug shell, which makes it really hard to
step through the test code.

This patch discards all output from spawned processes during test runs if
DISCARD_SUBPROCESS_OUTPUT environment variable is set (default for the
run-tests.sh script).

Bug: 1550365
Change-Id: I3762de6d3b1eac478b4a7ffbd66aa974f2dfdc86