update scheduler_status in reserve/unreserve methods instead of allocate/release
Previously, the SystemResource.allocate() and SystemResource.release()
methods were responsible for updating the System.scheduler_status
attribute, to indicate to whether the scheduler should look at the
system for scheduling.

However the SystemResource.release() method had the statement in the
wrong place. That method is called on *every* resource in *every* recipe
in a job, even if the resource has already been released. That means the
scheduler_status attribute was being incorrectly set to Pending in cases
where the resource had already been released before. The statement
should have gone after the short-circuit condition at the top of that
method.

This patch adds a comment on top of all the resource .release() methods
to point out this very non-obvious convention. There is already
a comment to this effect in Recipe.cleanup() but you don't see that when
modifying the .release() methods.

And this patch also moves the responsibility for updating
scheduler_status into the System.reserve() and System.unreserve()
methods, which seems like a cleaner place anyway. Those methods are only
called when a change is actually being made, which is what we want.

Bug: 1558776
Change-Id: Ic88a6a343d2fcca8e6e71a31dad444b28ea7518a