Allow to remove the keystone trust in Beaker
If a trust has been invalidated outside of Beaker (e.g. OpenStack deployment was
upgraded), it is possible that the VirtManager can not be instantiated due to
the session can not be established.

In that instance ignore the fact that we can not delete the trust in OpenStack
and just proceed with deleting it in Beaker. Otherwise the only way to remove
the old trust is by creating a new trust.

Bug: 1557847
Change-Id: I8d5ebc04ccbc4066aa0f0ab09cdd823bb67f2308