fix "not enough systems" logic
This regressed when it was refactored in commit a713fd08.

Previously, the "not enough systems" logic was in a separate beakerd
loop to the system allocation code (process_queued_recipe vs.
schedule_processed_recipe) which means there was an intervening session
flush and close between them, so that the changes to recipe.systems were
visible to the later queries.

Now that these steps all happen inside a single transaction, we need to
explicitly flush changes to recipe.systems before any queries use
recipe.dyn_systems.

Bug: 1558828
Change-Id: I671683a69e6a696be7d0c453c967406a95446fd7