Lookup iPXE compatible URLs to generate script
The ipxe-script end point relies on generating the script with HTTP/FTP URLs for
iPXE to boot from. If the primary URL is an NFS URL the joining of NFS urls with
the kernel and initrd paths will result in relative paths instead of URLs.

This partially puts in code which the end point used to lookup HTTP/FTP URLs
from the lab controller in order for iPXE to work.

Bug: 1563072
Change-Id: I0d94665f27d1da0b91380e0ade29b33f789f4103