Import separate RHEL8 composes
RHEL8 is being delivered as 3 composes instead of one: BaseOS the minimum system
and AppStream + Buildroot. The distro import tool understands to import a single
compose e.g. BaseOS, but has no idea about treating BaseOS, AppStream and
Buildroot as a single distro.

This patch adds additional functions and methods to make the distro import tool
aware of the modular approach RHEL8 takes. It uses a mixture of hard coded
values to add the additional repositories to RHEL8 so that Beaker includes the
additional repositories for RHEL8 provisioning.

Bug: 1554472
Change-Id: Ifdcdf06761f928727afd650dadc388db11a8e97e