avoid logging reporting database password to beakerd.log
Currently, if Graphite integration is configured, beakerd will load the
reports_engine.dburi setting if one is configured, and then print the
value of that setting in a debug message to /var/log/beaker/beakerd.log.
It will contain the database password, which is not ideal.

This changes the log message to use the sqlalchemy Engine object *after*
it's created. This correctly censors the password from the repr().

    2018-04-17 16:08:17,440 bkr.server.util DEBUG Created reports_engine Engine(mysql://dcallagh:***@solid.dcallagh.beakerdevs.lab.eng.bne.redhat.com/beaker_dcallagh?charset=utf8)

Change-Id: I901d2831d341a7cf1ffc04f556ed95c5a5e08d22