avoid calling reduced_install_options() in GuestRecipe.to_xml()
This call was introduced in Beaker 24.0, commit cee5f081, for bug
1410089, in order to correctly populate the location="" attribute for
guest recipes.

For new recipes (Beaker >= 25) we should just be using
recipe.installation.tree_url instead. That's what it's for, and it's
cheaper as well. Plus it will work correctly with custom <distro/> guest
recipes.

For old recipes (Beaker < 25) the reduced_install_options() method no
longer works, so we can't keep calling that. It's not really important
that location="" is still correct or consistent on a recipe that is
already completed (for example, the URLs could change in the distro
library which would also change the location="" attribute on old
recipes). But as a best-effort hack, we can look at just the recipe
ksmeta (instead of the full chain of install options) and pick a distro
tree URL from that. This should give the same answer in most cases. The
proper solution is to store the actual URL we used, which we do now in
recipe.installation.tree_url anyway.

Bug: 1568238
Change-Id: Idfbbd4f685ceb3b1ddd480af1e5092551314a794