Abort recipeset if one of the recipes fails to provision
If an exception was thrown during the provisioning of a recipe, the error would
have not bubbled up in order to abort the recipesets provisioning. This patch
makes sure that the exception is re-thrown in order for the recipesets
transaction to be rolled back and the recipesets status to be set to aborted.

Bug: 1568224
Change-Id: Id795c28292506922a176b931765c87d1f51b3452