Fix interface error
Our MySQL Python bindings throw an interface error if the underlying database
throws an error they don't know how to handle:

InterfaceError: (_mysql_exceptions.InterfaceError) (-1, 'error totally whack')
[SQL: 'INSERT INTO retention_tag (id, expire_in_days, needs_product) VALUES (1,
0, 0)']

The first interface error uses the same fix as in
https://gerrit.beaker-project.org/#/c/5951/.

The second interface error is related to the first, but since the current code
base does not have the attribute anymore, the patch addresses it in the
migration itself. The downgrade was therefore changed to adding the column as
non-nullable, which assigns a Boolean value for existing entries and then alter
the column to be NULLable.

Change-Id: I8487efa03f46bc7b6daf880615c4ffd1e842da96