beakerd: avoid holding open a database transaction in the outer loop
At the top of the outer loop functions beakerd was issuing a query
without any corresponding session.begin(), which means the connection
and its corresponding transaction was being left open until the next
invocation of a beakerd outer loop function.

If beakerd starts to get bogged down processing a large number of New
recipes, as in bug 1574772, this means that the transaction outside the
loop can be held open for many hours, which then makes everything worse
as InnoDB has to maintain a very long transaction history.

Change-Id: I27503a50735b2a7b9fa29fee2b96c4fee53499c8