beaker-log-delete: avoid holding open a database transaction outside the loop
Similar to commit 48cda168, beaker-log-delete may run for a long time if
it has a lot of jobs to purge. We want to avoid having an active
transaction for the entire duration of the run, because it puts extra
strain on Innodb and may contribute to the likelihood of bug 1574772
occurring.

Change-Id: Ifd0243eea18d57892d24098610b23d8aab8e31dd