Coerce score to int MySQL 5.x style
With the switch to MariaDB, the database is less tolerant of different types
than what is declared in the schema. The production instance is running in
strict mode, which enforces this schema. MySQL would simply coerce the string to
an integer depending on what you have passed in, MariaDB raises an error.

In order to keep backwards compatibility, we do in Python what MySQL did.

Bug: 1586049
Change-Id: I500db2e313f5563a75a7cf9cd6563cd5d090f6eb