Allow to import RHEL8 Alpha
This patch allows to import RHEL8 Alpha composes. The compose paths are hard
coded and any deploys which provide the compose should follow the hard coded
naming scheme of folders in order for the importer to import the compose.

Reasons for this are, that once available the name of the composes should and
will not change.

Change-Id: I43d7fa1749822cad98138d2b4d64916f6f0166d2