Always truncate old/new values for Activity model
When using MySQL, the db could handle being passed a User object and
would truncate it appropriately. There was prexisting code to truncate
the new/old values but it only expected unicode and was skipping User
objects (Which is what is passed to Activity when a system owner is
changed). As MariaDB is not so leanient this causes failures. This fix
will cast all objects passed in to unicode and then truncate them; as
was originally intended.

Bug: 1591391
Change-Id: I43c9601a83d042a28547475970e222f87ad1c9cc