delete unnecessary unicode_truncate() function
This was added way back in commit d917d76a because we observed that
MySQL appeared to be truncating string values in the middle of UTF-8
characters, leaving behind invalid bytes.

The *actual* cause of that was because Beaker was connecting to the
database with latin1 encoding. The column was 60 chars so MySQL
truncated the string at 60 *bytes*.

After we switched the database encoding to utf8 in commit f47cf808, this
was no longer an issue -- MySQL would correctly limit the field to 60
Unicode *characters* instead of 60 bytes.

Change-Id: I54fab1686106ab26743f69ee97aee80686eb8342