Handle RHEL8 synced to partners as well
The current distro import code will not be able to import RHEL8 Alpha composes
synced to partner repositories. What complicates this is a different directory
structure between partner repos and internal repos.

This patch adds support for importing RHEL8 alpha internally and for partners
adding code which tries to guess the directory layout.

Bug: 1593042
Change-Id: Ib9219d051b1b76cd5c7863a6cd012b9ed29d7d19