remove support for conditionally building the server bits in beaker.spec
This was originally needed for RHEL5, where we continued supporting and
building the client several years after we dropped support in the
server.

Nowadays, we are always building all subpackages everywhere, so we can
simplify the spec a little by removing these conditionals. They will
(hopefully) not be needed again in future.

Change-Id: I9787bb056e9b0f3dd9f17714528fb3b21e4de420