docs: describe /distribution/check-install
Also retain the description of /distribution/install as it has a long
life ahead of it yet.

Removed the suggestion that the install task uploads Anaconda logs,
because it actually didn't really upload many of them. Anamon already
does a much better job of this, and the Installation tab of the recipe
page is generally the right place to look first.

Bug: 1188539
Change-Id: Ia82c7780d75e2e7ed36ae27a69f01477703017ed