show lists in sorted order on the task page
These fields are lists of string identifiers. They come back from the
server in arbitrary database order. Let's sort them alphabetically so
they look nice for the humans.

This also helps the tests too. I found this because an unrelated patch
was causing a test failure, like this:

    ======================================================================
    FAIL: test_exclusive_releases (bkr.inttest.server.selenium.test_tasks.TestSubmitTask)
    ----------------------------------------------------------------------
    Traceback (most recent call last):
      File "/usr/lib/python2.7/site-packages/bkr/inttest/server/selenium/test_tasks.py", line 269, in test_exclusive_releases
	'RedHatEnterpriseLinux6\n'
    AssertionError: u'RedHatEnterpriseLinux7\nRedHatEnterpriseLinux6' != 'RedHatEnterpriseLinux6\nRedHatEnterpriseLinux7'

It was because my patch had caused RedHatEnterpriseLinux7 to be inserted
into the database earlier than RedHatEnterpriseLinux6, thus causing the
server to start returning the values in a different order.

Change-Id: I7e9c906c73420bf2f1feebdf057a9b92b9b9fe63