restructure beaker-watchdog using gevent, tidy up error handling
This adjusts the beaker-watchdog daemon to closely match the structure
of beaker-provision. The main loop is a polling loop which spawns and
manages a set of greenlets.

The actual aim of this patch is to clean up the error handling behaviour
and shutdown behaviour, so that it doesn't crash mysteriously if an
error occurs in any of the loops.

Bug: 991269
Change-Id: I064d4253c9a933601c2f6da2b5a3296dc746ad07