fill in custom packages into %packages in sorted order
This test was failing intermittently because the custom packages are
being filled into the kickstart in database order, which is arbitrary.

It would be nicer if we could put them into the kickstart in the same
order they appear in the job XML, but we do not persist the ordering so
the best we can do instead is obtain a consistent ordering by sorting
them on the way out.

This should fix the test, but it should also be nicer for humans too.
The packages will not appear in the kickstart in a totally
random-looking order.

Change-Id: Ia3d957232219c81d0272406643e4178a79bac87f