fix DECIMAL columns if their scale is wrong
The default MySQL precision and scale for DECIMAL (or Numeric in Python
land) is 10,0 but for unexplained reasons the Red Hat production
instance of Beaker has these columns as DECIMAL(10,2) instead.

This patch makes the precision and scale explicit for all such columns,
and adds extra checks to test_database_migration to catch
inconsistencies. A migration is added to fix any old Beaker instances
which have this problem.

Bug: 1600281
Change-Id: Ia7b94c40fca81adb2c0075f1caf384f9e1c64aa3