beaker-repo-update: handle missing OS majors as a special case
... instead of suppressing *all* exceptions when syncing the repo. In
case of an "expected" missing repo (such as Atomic Host) the warning
goes from this scary-looking message:

    2018-08-28 12:48:33,815 beaker-repo-update WARNING Cannot retrieve repository metadata (repomd.xml) for repository: http:--localhost:19998-BlueShoeLinux5-. Please verify its path and try again

to this much clearer message:

    2018-08-28 12:56:46,640 beaker-repo-update WARNING Harness packages not found for OS major BlueShoeLinux5, ignoring

And more importantly, any other error conditions (like a full disk or
a failure to fetch a particular package) will cause beaker-repo-update
to bail out immediately with a traceback. We want that, because
otherwise it is too easy to miss the real errors as they fly by buried
under subsequent debug messages.

Bug: 1619969
Change-Id: If972a87be5ed9b246b418ba5bcd08f31f273df64