Update Beaker Provided Tasks documentation
1. Update /distribution/command example

rhts-reboot does not check $REBOOTCOUNT so the current example creates
an infinite reboot loop.  Update the example to something safer like
logging the output of /proc/cpuinfo.

2. Replace "Other tasks" section with "Task source code"

The "Other tasks" section refers to the Tasks/ subdirectory which was
removed with commit 3b1960572abe in 2013.  Change this section to
instead link to the core-tasks and meta-tasks git repos.

Bug: 1699255
Change-Id: I50090abb49e4611813c9cd66856b4e1993f6c928