Add support for a 'ks_keyword' variable to 'ks_meta' attribute
Add a new 'ks_meta' variable 'ks_keyword'. This can be used to change
the default kickstart keyword from 'ks' to something else.

For example to support Debian can use:
  ks_meta="ks_keyword=preseed/url'

Then normal template processing will occur but instead of the kernel
command line argument being like:
    ks=http://example.com/kickstart/23

It would be:
    preseed/url=http://example/com/kickstart/23

If a user were to use this then an appropriate preseed template would
need to be put into /etc/beaker/kickstarts/ on the beaker server. For
example for Ubuntu would create the preseed Jinja2 file in
/etc/beaker/kickstarts/Ubuntu. For Debian it would be
/etc/beaker/kickstarts/Debian.

Bug: 1795917
Change-Id: I43a9b10be6122d51f164d689112530e2c9d474d1