Change authentication cookie to a persistent cookie
This is a quality of life improvement. Previous beaker used session
cookies. So when you close your browser the cookie would disappear and
then you would need to login again the next time you went to the
server.

Now we set a persistent cookie which has its expiration set based on
the "visit.timeout" value in the Beaker server.cfg file.

Fixes: #19