Don't set empty default values
when using restraint to run tests in DCI we set LOOKASIDE variable
from the job xml but couldn't figure out why it wasn't being set
when the job ran.  This default was overwriting our value.

We worked around this by creating a /etc/beaker/snippets/lab_env
file with nothing in it.