Fix: Get console logs raises Exception
When a guest recipe console log was being processed during beaker
upgrade, an exception was raised with call to startswith in
labcontroller.utils.get_console_files. In the case of guest recipes,
it is possible for the system name to be None. This change recognizes
this case, reports this occurred, and discontinues getting console
files for this guest recipes. Since exception is not raised, other recipes
will continue to get their console files.

Closes: BKR-5038