Fix: 500 internal error returned for some users
When getting users from beaker by way of
https:/beaker.com/users/username and that user is retired from company,
seeing a "500 internal error" instead of "Not Found" error.  The
failure is in get user LDAP look-up since an exception is raised when
looking for the "mail" key entry that no longer exists. Code
now checks whether it is present. If not, None returned for user.

Closes: BKR-5058