Replace old deprecated unittest method aliases
Those were removed in Python 3.12.

See https://docs.python.org/3.12/whatsnew/3.12.html#removed

Note that this was a search and replace job,
some of the asserts might be easier to read if spelled out differently,
e.g. using assertIn(..., ...) instead of assertTrue(... in ...),
but this is out of scope of this commit.