Fix curl help in fetch_wrapper for newer releases
Unable to retry fetch operations for latest Fedora releases
and RHEL-9+. These distros have an extra option
in the `curl help` in order to find the retry option help text.
Instead of just `curl --help` they use `curl --help all`.
Because help not working as expected, current code unable to
determine whether the `retry` option is available for `curl`
so retry not performed.